add AbstractBeanDefinition to avoid duplication
